backend/requirements.txt
fastapi
starlette
pydantic
uvicorn
python-multipart
pillow
openai